vaida/backend/requirements.txt
# VAIDA Backend - Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.18
httpx==0.28.1
openai==1.58.1
reportlab==4.2.5
jinja2==3.1.4

# Security
cryptography==44.0.0   # Fernet AES encryption for PII
slowapi==0.1.9         # Rate limiting middleware

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0

# Production extras (uncomment as needed)
# twilio==9.4.0                             # SMS OTP provider
# asyncpg==0.30.0                           # PostgreSQL async driver
# psycopg2-binary==2.9.10                   # PostgreSQL sync driver
# sentry-sdk[fastapi]==2.19.2               # Error monitoring
# prometheus-fastapi-instrumentator==7.0.2  # Metrics
